refactor: 불필요한 코드, 파일 구조 정리
