+++ index.html
@@ -4,21 +4,19 @@
     <meta charset="UTF-8" />
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-
     <link rel="preconnect" href="https://fonts.googleapis.com" />
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
     <link
       href="https://fonts.googleapis.com/css2?family=Noto+Sans:wght@400;700&display=swap"
       rel="stylesheet"
     />
-
-    <link rel="stylesheet" href="style.css" />
     <title>홈 | 카카오페이지</title>
     <link
       rel="shortcut icon"
       href="https://static-page.kakao.com/static/common/favicon.ico?b65d26a018e7e801d042589459ab3611"
       type="image/x-icon"
     />
+    <link rel="stylesheet" href="style.css" />
     <script src="main.js" type="module"></script>
   </head>
   <body>
