--- scss/_badges.scss
@@ -1,8 +1,0 @@
-.noti--cnt {
-  padding: 4px 6px;
-  margin-left: 5px;
-  font-size: 13px;
-  background-color: $yellow;
-  border: 0;
-  border-radius: 10px;
-}
