+++ scss/_main.scss
@@ -141,6 +141,15 @@
   }
 }
 
+.noti--cnt {
+  padding: 4px 6px;
+  margin-left: 5px;
+  font-size: 13px;
+  background-color: $yellow;
+  border: 0;
+  border-radius: 10px;
+}
+
 .webtoon-menu__list--top-left {
   border-radius: 5px 0 0 0;
 }
