+++ scss/style.scss
@@ -1,5 +1,4 @@
 @import 'header';
 @import 'main';
 @import 'footer';
-@import 'badges';
 @import 'utility';
